Updating for new exp decay function and incentive weighting (#501)
* Updating for new exp decay function and incentive weighting

* update volatility

* Updating miner statistics and config

* Updating minerstatistics for new penalized and augmented params

* Updating docs

* ruff and tests

* Updating terminology and some metrics

* skip ChallengesProps when not in challengeperiod

* update dashboard

* Updating some bugs in the calculation of the penalties and weights for the metrics

* update challengeperiod scorecards

* Updating for new ledger penalties

* Tweaking scoring categories

* update dash drawdown

* add cumulative return to ledger checkpoints

* fix PydanticSerializationError for PerfCheckpoint

* update readme for mdd elimination

* Tweaking scoring to clean up and double checking against prior calculations

* Updating tests and notations on functions in metrics

* Updating stats metrics for weighting on tstat

* calmar ratio text

* Log dash data generation time

* Allow quote sources to update independently of price sources being updated

* Reverting adjusted annual returns

---------

+++ vali_objects/utils/functional_utils.py
@@ -17,4 +17,4 @@ class FunctionalUtils:
             raise ValueError("The spread parameter must be different from 0")
 
         exp_term = np.clip(spread * (x - shift), -100, 100)
-        return np.clip(1 / (1 + np.exp(exp_term)), 0, 1)
+        return float(np.clip(1 / (1 + np.exp(exp_term)), 0, 1))
